requirements.txt
# Core LLM & RAG (env 가상환경 설치 버전)
langchain-aws==1.1.0
langchain-community==0.4.1
langchain-core==1.2.5
langgraph==1.0.5

# Vector Database
faiss-cpu==1.13.2

# AWS
boto3==1.42.17

# Web Framework
fastapi==0.128.0
uvicorn==0.40.0
streamlit==1.52.2

# Utilities
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5

# mcp
#mcp
